Fix placeholder image and environment
